+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTRBCDVZ3T3NCW3PQ1PFZX25"
+        "01KTRJBCC3069625PE43J9MN9X"
     )
     r"""The unique identifier of the data source"""